refactor(phase1c): library.dataset を train_util.py から分離
データセット骨格となる以下の要素を新規モジュール library/dataset.py に移動。
train_util.py 側は後方互換のため re-export shim を残す。

PR-1（dataset 系）の 3 つ目の PR。約 1,800 行の最大級の抽出。

## 移動した要素（library/dataset.py、約 1,960 行）

定数:
- IMAGE_EXTENSIONS（pillow_avif / jxl の条件付き拡張含む）
- IMAGE_TRANSFORMS
- TEXT_ENCODER_OUTPUTS_CACHE_SUFFIX / _SD3

クラス:
- ImageInfo（画像メタデータ）
- BucketManager（アスペクト比 bucket）
- BucketBatchIndex（NamedTuple）
- AugHelper（color augmentation）
- BaseDataset（DreamBooth/FineTuning/ControlNet 共通の抽象データセット、約 1,260 行）
- DatasetGroup（torch.utils.data.ConcatDataset サブクラス）
- MinimalDataset（ユーザー定義データセット用）

関数:
- split_train_val
- debug_dataset
- glob_images / glob_images_pathlib
- load_arbitrary_dataset
- load_image

## 後方互換

DreamBoothDataset / FineTuningDataset / ControlNetDataset の 3 つは引き続き
train_util.py に残る（PR-1d で移動予定）。BaseDataset を継承する関係で、
train_util 内のサブクラスは re-export 経由で BaseDataset を参照する。

## 循環依存の整理

caching.py を更新:
- _ds() ヘルパーを追加し、library.dataset の BucketManager / load_image /
  IMAGE_TRANSFORMS を遅延参照
- _tu() は HIGH_VRAM / get_hidden_states_sdxl 用に維持
- caching.py 内の重複 IMAGE_TRANSFORMS 定義を削除（library.dataset を単一の
  ソースに）

## 内部参照の切替

- tests/test_validation.py: from library.dataset import split_train_val
- finetune/prepare_buckets_latents.py: dataset.BucketManager / dataset.ImageInfo を直接使用

学習スクリプト（train_*.py、fine_tune.py 等）の train_util.X 属性アクセスは
shim 経由で動作するため、後続 PR で順次切替予定。

## 動作確認

- ✅ library.dataset / caching / train_util / config_util / strategy_base 単独 import
- ✅ 両方向の import 順序
- ✅ 全主要学習スクリプト（fine_tune, train_db, train_network, sdxl/sd3/flux/lumina/anima 系）の py_compile
